+++ src/js/src/utils/disk.js
@@ -1,0 +1,3 @@
+export function getDiskLabel(disk) {
+  return disk.displayName || disk.name || disk.fullPath;
+}